Fix Python 3.10 asyncio.gather exception
In Py 3.10 `asyncio.gather` wont' take the loop in as a kwarg, but rather use `_events._get_event_loop()` which, since there's no loop set, will raise an exception.
But this only happens if there are no tasks to be cancelled. If there are everything is going to be fine.
For simplicity I put the whole thing under the `if` statement, because running `asyncio.gather` with no tasks would have no effect on older versions either.

I suggest adding a regression test for this, at some point.
Just please merge it ASAP, because I don't like having modified vendor code :D

Also, thanks for the awesome lib, I'm thankful I got rid of the boilerplate.